Robinhood/backend/requirements.txt
Flask==2.0.2
yfinance==0.2.40
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.27
Werkzeug==2.2.2
Flask-CORS==3.0.10
flask_socketio==5.3.6
openai==0.28
python-dotenv==1.0.1
